+++ src/commonMain/kotlin/uk/gov/hmrc/calculator/utils/clarification/Clarification.kt
@@ -16,23 +16,30 @@
 package uk.gov.hmrc.calculator.utils.clarification
 
 // Description added for readability clarifying what each Clarification enum represent, no real usage in code yet.
-enum class Clarification(private val description: String) {
-    NO_TAX_CODE_SUPPLIED("Did not provide a tax code."),
-    HAVE_STATE_PENSION("Have state pension."),
-    HAVE_NO_STATE_PENSION("Have no state pension."),
-    SCOTTISH_INCOME_APPLIED("Applied Scottish Income Tax rates."),
-    SCOTTISH_CODE_BUT_OTHER_RATE("Applied Scottish Income Tax rates because you used a Scottish tax code."),
-    NON_SCOTTISH_CODE_BUT_SCOTTISH_RATE("Applied Scottish Income Tax rates as you said you pay Scottish Income Tax."),
-    INCOME_BELOW_STUDENT_AND_POSTGRAD_LOAN("Income is below the level needed to start repaying student/postgrad loan."),
-    INCOME_BELOW_STUDENT_LOAN("Income is below the level needed to start repaying student loan."),
-    INCOME_BELOW_POSTGRAD_LOAN("Income is below the level needed to start repaying postgrad loan."),
+enum class Clarification(val priority: Int, private val description: String) {
+    HAVE_STATE_PENSION(1, "Have state pension."),
+    HAVE_NO_STATE_PENSION(2, "Have no state pension."),
+    INCOME_OVER_100K_WITH_TAPERING(3, "Income over £100K with default 1257L tax code, tapering applied."),
+    NO_TAX_CODE_SUPPLIED(4, "Did not provide a tax code."),
+    SCOTTISH_INCOME_APPLIED(5, "Applied Scottish Income Tax rates."),
+    SCOTTISH_CODE_BUT_OTHER_RATE(6, "Applied Scottish Income Tax rates because you used a Scottish tax code."),
+    NON_SCOTTISH_CODE_BUT_SCOTTISH_RATE(
+        7,
+        "Applied Scottish Income Tax rates as you said you pay Scottish Income Tax."
+    ),
+    K_CODE(8, "Tax code started with K."),
+    PENSION_EXCEED_ANNUAL_ALLOWANCE(9, "Pension contribution amount is higher than annual allowance."),
+    PENSION_BELOW_ANNUAL_ALLOWANCE(10, "Pension contribution amount is below than annual allowance."),
+    INCOME_BELOW_STUDENT_AND_POSTGRAD_LOAN(
+        11,
+        "Income is below the level needed to start repaying student/postgrad loan."
+    ),
+    INCOME_BELOW_STUDENT_LOAN(12, "Income is below the level needed to start repaying student loan."),
+    INCOME_BELOW_POSTGRAD_LOAN(13, "Income is below the level needed to start repaying postgrad loan."),
     INCOME_BELOW_STUDENT_BUT_ABOVE_POSTGRAD_LOAN(
+        14,
         "Income is below the level needed to start repaying student loan, but high enough to repay postgrad loan"
     ),
-    PENSION_EXCEED_INCOME("Pension contribution amount is higher than income."),
-    PENSION_EXCEED_ANNUAL_ALLOWANCE("Pension contribution amount is higher than annual allowance."),
-    PENSION_BELOW_ANNUAL_ALLOWANCE("Pension contribution amount is below than annual allowance."),
-    INCOME_OVER_100K("Income over £100k"),
-    INCOME_OVER_100K_WITH_TAPERING("Income over £100K with default 1257L tax code."),
-    K_CODE("Tax code started with K.")
+    INCOME_OVER_100K(15, "Income over £100k with user specified 1257L tax code, no tapering."),
+    PENSION_EXCEED_INCOME(16, "Pension contribution amount is higher than income."),
 }